ActionDataGrid deleteItem event fix (#448)
Only fire deleteItem event in SingleActionDataGrid when there is a selected item

+++ src/components/ui/DataGrid/SingleActionDataGrid.svelte
@@ -78,7 +78,7 @@
   }
 
   function onKeyDown(event: KeyboardEvent) {
-    if (isDeleteEvent(event)) {
+    if (selectedItemId !== null && isDeleteEvent(event)) {
       deleteItem();
     }
   }
